WIP: change mapping of uint{8,16,32,64}
This changes the mapping of uint types to non-widening yang.common
classes and deals with them being used by yang.{data,model}.api:

YANG type    Old Java type       New Java Type
uint8        java.lang.Short     org.opendaylight.yangtools.yang.common.Uint8
uint16       java.lang.Integer   org.opendaylight.yangtools.yang.common.Uint16
uint32       java.lang.Long      org.opendaylight.yangtools.yang.common.Uint32
uint64       java.math.BigIteger org.opendaylight.yangtools.yang.common.Uint64

This affects all getters, setters and value wrapper classes. For
setters, code generator emits a deprecated setter, which performs
the conversion internally.

JIRA: MDSAL-330
JIRA: YANGTOOLS-1018

+++ binding/mdsal-binding-generator-impl/src/test/java/org/opendaylight/mdsal/binding/generator/impl/GenerateInnerClassForBitsAndUnionInLeavesTest.java
@@ -102,7 +102,7 @@ public class GenerateInnerClassForBitsAndUnionInLeavesTest {
                                         }
                                     } else if (unionLeafProperty.getName().equals("uint8")) {
                                         uint8UnionPropertyFound = true;
-                                        if (unionLeafPropertyType.equals("Short")) {
+                                        if (unionLeafPropertyType.equals("Uint8")) {
                                             uint8UnionPropertyTypeOK = true;
                                         }
                                     }
